Add webpack. Split markup functions into modules
