+++ public/index.html
@@ -32,6 +32,6 @@
           </section>
       </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
