+++ src/components/menu.js
@@ -1,0 +1,13 @@
+/**
+ * Get HTML string template for menu
+ *
+ * @return {string}
+ */
+export const getMenuTemplate = () => {
+  return (
+    `<nav class="trip-controls__trip-tabs  trip-tabs">
+      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
+      <a class="trip-tabs__btn" href="#">Stats</a>
+    </nav>`
+  );
+};
